// types/profile.ts
export interface StudentProfile {
    email: string;
    firstName: string;
    middleName?: string;
    lastName: string;
    extensionName?: string;
    birthDate: Date | undefined;
    gender: string;
    mobileNumber: string;
    isWithLrn: boolean;
    isReturning: boolean;
    schoolYear: number;
    gradeLevel: string;
    birthPlace: string;
    motherTongue: string;
    birthCertificate: string;
    learnerReference: string;
    isWithCommunity: boolean;
    community?: string;
    isBeneficiary: boolean;
    beneficiary?: string;
    currentHouseNumber: string;
    currentStreet: string;
    currentBarangay: string;
    currentCity: string;
    currentProvince: string;
    currentCountry: string;
    permanentHouseNumber: string;
    permanentStreet: string;
    permanentBarangay: string;
    permanentCity: string;
    permanentProvince: string;
    permanentCountry: string;
    fatherFirstName: string;
    fatherMiddleName?: string;
    fatherLastName: string;
    fatherExtensionName?: string;
    motherFirstName: string;
    motherMiddleName?: string;
    motherLastName: string;
    motherExtensionName?: string;
    guardianFirstName: string;
    guardianMiddleName?: string;
    guardianLastName: string;
    guardianExtensionName?: string;
    yearCompleted: number;
    gradeCompleted: string;
    lastSchoolAttended: string;
    lastSchoolId: string;
    track?: string;
    strand?: string;
    isFirstSemester: boolean;
    modalities?: string[];
    role: string;
}